pyproject.toml
[project]
name = "langgraph-routing-logic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grandalf>=0.8",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.5",
    "langgraph>=0.5.1",
    "langgraph-cli[inmem]>=0.3.3",
    "python-dotenv>=1.1.1",
]
